Pull out JukeBox implementation from framework.cpp
Shouldn't have any functional changes
