+++ game/ui/general/mainmenu.cpp
@@ -4,8 +4,8 @@
 #include "forms/ui.h"
 #include "framework/event.h"
 #include "framework/framework.h"
+#include "framework/jukebox.h"
 #include "framework/keycodes.h"
-#include "framework/sound.h"
 #include "game/ui/debugtools/debugmenu.h"
 #include "game/ui/general/difficultymenu.h"
 #include "game/ui/general/loadingscreen.h"
