+++ game/ui/tileview/cityview.cpp
@@ -15,6 +15,7 @@
 #include "framework/font.h"
 #include "framework/framework.h"
 #include "framework/image.h"
+#include "framework/jukebox.h"
 #include "framework/keycodes.h"
 #include "framework/renderer.h"
 #include "framework/sound.h"
